syscontainers: use temporarily CLI ostree for checkouts
This is required to process Docker whiteouts files in OSTree, which are
not enabled by this patch.

There is an issue in the way the RepoCheckoutOptions object is mapped by
glib, as the C struct is using bit fields that are not supported by the
introspection, so use the CLI version of OSTree for now.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>