syscontainers: add generation of an rpm
Add a hidden (for now) option ---generate-rpm to "install --system" to
generate an rpm file instead of installing the container.  The rpm can
be installed on the system and the container won't be managed by
atomic, update and uninstall are disabled.

Differently from containers handled by atomic that installs the
container under /var/lib/containers/atomic, the RPM version installs
its files under /usr/lib/containers and can be integrated in an atomic
image as a normal rpm.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>