syscontainers: support no-container-service mode
with the rpm generation, it can be useful to copy files to the host
without the overhead of maintaining a dummy service.

Containers that do not use a service, can specify it in the
manifest file as:

    "noContainerService": true

it also implies that the rootfs is deleted once all the files are
copied to the system.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>