syscontainers: introduce containers storage
Support the extraction of layers under $CHECKOUT/.storage/$ID so that
they can be used directly by overlayfs2.

This won't be used by checked containers as their rootfs must be ready
when the system starts and we don't want any setup code before the
container can run.

Differently, we have already two use cases that can benefit of this
feature and that are initialized directly from atomic:

1) mount of images
2) one shot containers

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>