syscontainers: move storage to /var/lib/containers/atomic when no AH
When /var/lib/containers/atomic is not on the same file system as
/ostree we weren't able to create hardlinks thus falling back to
copying files.  Additionally, when not running on Atomic Host, this
avoids to create a system wide /ostree that is used only for system
containers.

This problem doesn't exist on Atomic Host since we use the OSTree
sysroot to get the destination path for the checkouts.  On AH, we will
still continue to use /ostree/repo so that we are able to share the
storage with the system.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1495588

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>