syscontainers: fix detection of repo location
Improve the detection of the repo being on the same file system as the
checkout path by trying to link a file.  This is necessary since on a
bind mount the st_dev for the two directories is the same, but hardlinks
will still fail with EXDEV.

Also, change the test in _canonicalize_location to check that on an
Atomic Host /sysroot/ostree/deploy/*/var/* and /var/* have the same
st_dev and st_ino.  It is necessary so that files created under /sysroot
are visible in /var, and in this case it is fine the two locations are
on different file systems.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>