syscontainers: expose ATOMIC
atomic run has some more features compared to a raw "$RUNTIME exec",
like the possibility to run a command for a service which is not started
yet.  Expose $ATOMIC as well during the installation so that a container
can more easily build a wrapper script for files inside the container.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>